requirements-test.txt
# Testing dependencies for YOLO training repository

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # For parallel test execution

# Test utilities
pytest-html>=3.1.0  # HTML test reports
pytest-json-report>=1.5.0  # JSON test reports
pytest-benchmark>=4.0.0  # Performance benchmarking
pytest-timeout>=2.1.0  # Test timeout handling

# Mocking and patching
unittest-mock>=1.0.0
freezegun>=1.2.0  # Time mocking

# Code quality and linting
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0

# Documentation testing
doctest-ignore-unicode>=0.1.2

# Performance testing
pytest-profiling>=1.7.0

# Memory testing
pytest-memray>=0.3.0

# Coverage reporting
coverage>=6.0.0

# Test data generation
faker>=18.0.0
factory-boy>=3.2.0

# Image processing for tests
Pillow>=9.0.0
opencv-python>=4.5.0

# Scientific computing for tests
numpy>=1.21.0
pandas>=1.3.0

# Machine learning for tests
torch>=1.13.0
torchvision>=0.14.0

# YAML handling for tests
PyYAML>=6.0

# Environment management
python-dotenv>=0.19.0

# Progress bars for tests
tqdm>=4.60.0

# Additional utilities
requests-mock>=1.10.0  # HTTP request mocking
responses>=0.20.0  # HTTP response mocking
